Fix memory leak in spl utility
Detected by the clang static analyzer, a memory leak was possible
in the case when a realloc(3) failed.  In practice this very unlikely
to happen and even if it did the utility is short lived so it wouldn't
be a problem.  Nevertheless this analysis is correct as this patch
resolves the leak.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #400