Add hooks to memory allocate function for direct reclaim debugging
In order to better catch deadlocks related to direct reclaim, this patch add
hooks to kmem_alloc, kmem_cache_alloc and vmem_alloc in order to simulate
direct reclaim as we desire.

Before using it, you would need to register a callback function. The function
should in turn call to the target shrinker functions.

The whole code is covered with #ifdef KMEM_DEBUG_DIRECT_RECLAIM, so you need
to add it to spl_config.h to enable it.

Signed-off-by: Chunwei Chen <tuxoko@gmail.com>