Use raw_smp_processor_id instead of smp_processor_id
This means callers do not need to turn off preemption in places where
using this information is safe. Consequently, there is no need to modify
consumers to call `kpreempt_disable()` and `kpreempt_enable()`, which
means that we do not insert a `preempt_schedule()` into places where
there is none on Illumos.

Signed-off-by: Richard Yao <ryao@gentoo.org>