Increase spl_kmem_alloc_warn limit
In order to support ABD with large blocks the spl_kmem_alloc_warn
limits needs to be increased to 64K.

A 16M block requires that a pointers be stored for 4096 4K-pages
on an x86_64 system.  Each of these points is 8 bytes requiring
an allocation of 8*4096=32,768 bytes.  The addition of a small
header to this structure pushes the allocation over the default
32K warning threshold.

In addition, fix a small bug when MAX was used instead of MIN
when setting the default.  This ensures a reasonable limit is
still set on systems with page sizes larger then 4K.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>